made condition more precised, if table exists in the database and has data, then do remove old records

+++ src/Commands/DeleteRecordsBeforeCertainDays.php
@@ -4,8 +4,9 @@ namespace OTIFSolutions\LaravelTracker\Commands;
 
 use Carbon\Carbon;
 use Illuminate\Console\Command;
+use Illuminate\Support\Facades\Schema;
 use OTIFSolutions\Laravel\Settings\Models\Setting;
-use OTIFSolutions\LaravelTracker\Models\{UserSession, RequestData, UserActivity, MiscData};
+use OTIFSolutions\LaravelTracker\Models\{MyCookie, UserSession, RequestData, UserActivity, MiscData};
 
 class DeleteRecordsBeforeCertainDays extends Command {
 
@@ -17,27 +18,33 @@ class DeleteRecordsBeforeCertainDays extends Command {
 
         $keepDataExceptDays = Setting::get('keep_except') ?? Setting::set('keep_except', 10);
 
-        if (UserSession::exists()) {
+        if (Schema::hasTable('user_sessions') && UserSession::exists()) {
             UserSession::whereDate('created_at', '<=', Carbon::now()->subDays($keepDataExceptDays))->delete();
             $this->info('Users data before ' . $keepDataExceptDays . ' days removed');
             $this->newLine();
         }
 
-        if (UserActivity::exists()) {
+        if (Schema::hasTable('user_activities') && UserActivity::exists()) {
             UserActivity::whereDate('created_at', '<=', Carbon::now()->subDays($keepDataExceptDays))->delete();
             $this->info('Users activities data before ' . $keepDataExceptDays . ' days removed');
             $this->newLine();
         }
 
-        if (RequestData::exists()) {
+        if (Schema::hasTable('request_data') && RequestData::exists()) {
             RequestData::whereDate('created_at', '<=', Carbon::now()->subDays($keepDataExceptDays))->delete();
             $this->info('Users request data before ' . $keepDataExceptDays . ' days removed');
             $this->newLine();
         }
 
-        if (MiscData::exists()) {
+        if (Schema::hasTable('misc_data') && MiscData::exists()) {
             RequestData::whereDate('created_at', '<=', Carbon::now()->subDays($keepDataExceptDays))->delete();
-            $this->info('Users request data before ' . $keepDataExceptDays . ' days removed');
+            $this->info('Users miscellaneous data before ' . $keepDataExceptDays . ' days removed');
+            $this->newLine();
+        }
+
+        if (Schema::hasTable('my_cookies') && MyCookie::exists()) {
+            RequestData::whereDate('created_at', '<=', Carbon::now()->subDays($keepDataExceptDays))->delete();
+            $this->info('Users cookies data before ' . $keepDataExceptDays . ' days removed');
             $this->newLine();
         }
